Update 10. Smart_Kitchen.ino

+++ examples/NodeMCU-32S/Projects/10. Smart_Kitchen/10. Smart_Kitchen.ino
@@ -161,7 +161,7 @@ void loop() {
 
     //Publish telemetry data 2
     float h = dht.readHumidity();
-    float t = dht.readTemperature();
+    int t = dht.readTemperature();
 
     JSONVar payloadObject;
     payloadObject["Humidity"] = h;
